+++ login-workflow/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   Componentize `BrandedCardContainer` and added documentation.
+-   A `BrandedCardContainer` component.
 -   Confirm password field error state message in `ChangePasswordForm`.
 
 ## v3.1.1 (December 14, 2022)